move test_core_v3.py functions back into test_core.py
